refactor(Balance Sheet Report) - (test) Update E2E test for Account Reference - Split longest file
