+++ server/controllers/finance/reports/ohada_balance_sheet/index.js
@@ -18,6 +18,7 @@ const db = require('../../../../lib/db');
 const AccountReference = require('../../accounts/references');
 const ReportManager = require('../../../../lib/ReportManager');
 const conditionalReferences = require('../../accounts/conditionalReferences');
+const balanceSheetElement = require('./balanceSheetElement');
 
 // report template
 const TEMPLATE = './server/controllers/finance/reports/ohada_balance_sheet/report.handlebars';
@@ -30,72 +31,9 @@ const DEFAULT_PARAMS = {
   footerRight : '[page] / [toPage]',
 };
 
-const balanceSheetAssetTable = [
-  { ref : 'AD', is_title : 1 },
-  { ref : 'AE', is_title : 0 },
-  { ref : 'AF', is_title : 0 },
-  { ref : 'AG', is_title : 0 },
-  { ref : 'AH', is_title : 0 },
-  { ref : 'AI', is_title : 1 },
-  { ref : 'AJ', is_title : 0 },
-  { ref : 'AK', is_title : 0 },
-  { ref : 'AL', is_title : 0 },
-  { ref : 'AM', is_title : 0 },
-  { ref : 'AN', is_title : 0 },
-  { ref : 'AP', is_title : 0 },
-  { ref : 'AQ', is_title : 1 },
-  { ref : 'AR', is_title : 0 },
-  { ref : 'AS', is_title : 0 },
-  { ref : 'AZ', is_title : 1 },
-  { ref : 'BA', is_title : 0 },
-  { ref : 'BB', is_title : 0 },
-  { ref : 'BG', is_title : 0 },
-  { ref : 'BH', is_title : 0 },
-  { ref : 'BI', is_title : 0 },
-  { ref : 'BJ', is_title : 0 },
-  { ref : 'BK', is_title : 1 },
-  { ref : 'BQ', is_title : 0 },
-  { ref : 'BR', is_title : 0 },
-  { ref : 'BS', is_title : 0 },
-  { ref : 'BT', is_title : 1 },
-  { ref : 'BU', is_title : 0 },
-  { ref : 'BZ', is_title : 1 },
-];
+const balanceSheetAssetTable = balanceSheetElement.balanceSheetAssetTable();
 
-const balanceSheetLiabilityTable = [
-  { ref : 'CA', is_title : 1 },
-  { ref : 'CB', is_title : 0 },
-  { ref : 'CC', is_title : 0 },
-  { ref : 'CD', is_title : 0 },
-  { ref : 'CE', is_title : 0 },
-  { ref : 'CF', is_title : 0 },
-  { ref : 'CG', is_title : 0 },
-  { ref : 'CH', is_title : 0 },
-  { ref : 'CI', is_title : 1 },
-  { ref : 'CK', is_title : 1 },
-  { ref : 'CL', is_title : 0 },
-  { ref : 'CM', is_title : 0 },
-  { ref : 'CP', is_title : 1 },
-  { ref : 'DA', is_title : 0 },
-  { ref : 'DB', is_title : 0 },
-  { ref : 'DC', is_title : 0 },
-  { ref : 'DD', is_title : 0 },
-  { ref : 'DF', is_title : 1 },
-  { ref : 'DG', is_title : 1 },
-  { ref : 'DH', is_title : 0 },
-  { ref : 'DI', is_title : 0 },
-  { ref : 'DJ', is_title : 0 },
-  { ref : 'DK', is_title : 0 },
-  { ref : 'DM', is_title : 0 },
-  { ref : 'DN', is_title : 0 },
-  { ref : 'DP', is_title : 0 },
-  { ref : 'DQ', is_title : 0 },
-  { ref : 'DR', is_title : 0 },
-  { ref : 'DS', is_title : 0 },
-  { ref : 'DT', is_title : 1 },
-  { ref : 'DV', is_title : 1 },
-  { ref : 'DZ', is_title : 1 },
-];
+const balanceSheetLiabilityTable = balanceSheetElement.balanceSheetLiabilityTable();
 
 // expose to the API
 exports.document = document;
@@ -118,7 +56,7 @@ function document(req, res, next) {
     return;
   }
 
-  getFiscalYearDetails(params.fiscal_id)
+  balanceSheetElement.getFiscalYearDetails(params.fiscal_id)
     .then(fiscalYear => {
       _.merge(context, { fiscalYear });
 
@@ -174,8 +112,8 @@ function document(req, res, next) {
       }
 
       let list = [];
-      const currentReferences = formatReferences(_.groupBy(currentData, 'abbr'));
-      const previousReferences = formatReferences(_.groupBy(previousData, 'abbr'));
+      const currentReferences = balanceSheetElement.formatReferences(_.groupBy(currentData, 'abbr'));
+      const previousReferences = balanceSheetElement.formatReferences(_.groupBy(previousData, 'abbr'));
 
       const assetTable = balanceSheetAssetTable.map(item => {
         item.label = 'REPORT.OHADA.REF_DESCRIPTION.'.concat(item.ref);
@@ -296,54 +234,4 @@ function aggregateReferences(references, currentDb, previousDb) {
   });
 
   return item;
-}
-
-function formatReferences(references) {
-  const values = {};
-  _.forEach(references, (reference, key) => {
-    const [brut] = reference.filter(elt => elt.is_amo_dep === 0);
-    let [amortissement] = reference.filter(elt => elt.is_amo_dep === 1);
-
-    if (!amortissement) {
-      amortissement = { balance : 0 };
-    }
-
-    const net = {
-      abbr : brut.abbr,
-      description : brut.description,
-      // reduce amortissement from brut
-      // the amortissement is supposed to be < 0
-      // that the reason we use brut + amortissement which is implicitly brut - amortissement
-      balance : brut.balance + amortissement.balance,
-    };
-
-    values[key] = { brut, amortissement, net };
-  });
-  return values;
-}
-
-function getFiscalYearDetails(fiscalYearId) {
-  const bundle = {};
-  // get fiscal year details and the last period id of the fiscal year
-  const query = `
-    SELECT 
-      p.id AS period_id, p.end_date,
-      fy.id, fy.label, fy.previous_fiscal_year_id 
-    FROM fiscal_year fy 
-    JOIN period p ON p.fiscal_year_id = fy.id 
-      AND p.number = (SELECT MAX(period.number) FROM period WHERE period.fiscal_year_id = ?)
-    WHERE fy.id = ?;
-  `;
-  return db.one(query, [fiscalYearId, fiscalYearId])
-    .then(fiscalYear => {
-      bundle.current = fiscalYear;
-
-      return bundle.current.previous_fiscal_year_id
-        ? db.one(query, [bundle.current.previous_fiscal_year_id, bundle.current.previous_fiscal_year_id]) : {};
-    })
-    .then(previousFiscalYear => {
-      bundle.previous = previousFiscalYear;
-
-      return bundle;
-    });
 }
